try-catch for save then move

+++ trainer/im_utils.py
@@ -223,8 +223,27 @@ def save_then_move(out_path, seg_alpha):
     fname = os.path.basename(out_path)
     temp_path = os.path.join('/tmp', fname)
     imsave(temp_path, seg_alpha)
-    shutil.copy(temp_path, out_path)
-    os.remove(temp_path)
+
+
+    # this is crashing on colab. The file doesn't exist.
+    # keep trying until it exists.
+    copied = False 
+    limit = 30 # 3 seconds max
+    tries = 0
+    while not copied and tries < limit:
+        try:
+            shutil.copy(temp_path, out_path)
+            os.remove(temp_path)
+            copied = True
+        except Exception as _
+            time.sleep(0.1)
+    # if it still didn't work
+    if not copied:
+        # try one last time and this time no try-catch.
+        # just show the exception in console output.
+        shutil.copy(temp_path, out_path)
+        os.remove(temp_path)
+
 
 def load_image(photo_path):
     photo = imread(photo_path)
